Some changes - Removed individual declarations for width and height - Refactored lobbyMobile and lobbyDesktop to separate files - Renamed css classes with double unscores
