+++ packages/frontend/pages/lobby.tsx
@@ -1,41 +1,22 @@
-import { ReactElement, useState, useRef } from 'react';
+import { ReactElement } from 'react';
 import { connect } from 'react-redux';
-import { Grid } from 'semantic-ui-react';
-import Link from 'next/link';
 
 import { RootState } from '../store/index';
-import { ThemeOptions, IUserOptionsState } from '../store/userOptions/types';
-import {
-  MobileView,
-  ISystemState,
-  WindowDimensions,
-} from '../store/system/types';
+import { MobileView } from '../store/system/types';
 
-import Nav from '../components/nav';
-import OnlinePlayers from '../components/lobby/onlinePlayers';
-import Taako from '../components/lobby/taako';
-import Chat, { dateGenObj } from '../components/lobby/chat';
-import GamesMenu from '../components/lobby/gamesMenu/gamesMenu';
-import Announcements from '../components/lobby/announcements';
+import LobbyMobile from '../components/lobby/lobbyMobile';
+import LobbyDesktop from '../components/lobby/lobbyDesktop';
 
 interface IStateProps {
-  theme: ThemeOptions;
   mobileView: MobileView;
-  windowDimensions: WindowDimensions;
 }
 
-const Lobby = ({
-  mobileView,
-  theme,
-  windowDimensions,
-}: IStateProps): ReactElement => (
+type Props = IStateProps;
+
+const Lobby = ({ mobileView }: Props): ReactElement => (
   <>
     <title>Lobby</title>
-    {mobileView ? (
-      <LobbyMobile theme={theme} windowDimensions={windowDimensions} />
-    ) : (
-      <LobbyDesktop />
-    )}
+    {mobileView ? <LobbyMobile /> : <LobbyDesktop />}
     <style global jsx>
       {`
         // CSS to make NextJS Page one page tall
@@ -76,284 +57,8 @@ const Lobby = ({
   </>
 );
 
-const LobbyInfo = (): ReactElement => (
-  <>
-    <Grid.Row className="taako_wrapper">
-      <Taako />
-    </Grid.Row>
-    <Grid.Row>
-      <Announcements
-        announcements={[
-          {
-            id: '1',
-            timestamp: new Date(dateGenObj.next().value as number),
-            link: '/announcements/123',
-            text: 'New Patreon rewards!',
-          },
-          {
-            id: '2',
-            timestamp: new Date(dateGenObj.next().value as number),
-            link: '/announcements/123',
-            text: 'Morgana cannot lie anymore!',
-          },
-          {
-            id: '3',
-            timestamp: new Date(dateGenObj.next().value as number),
-            link: '/announcements/123',
-            text: 'Players will now be allowed to wear bows',
-          },
-        ]}
-      />
-    </Grid.Row>
-    <Grid.Row>
-      <OnlinePlayers
-        players={[
-          { username: 'ProNub' },
-          { username: 'Skies' },
-          { username: 'Pam' },
-          { username: 'Pam2' },
-          { username: 'Pam3' },
-          { username: 'Pam4' },
-          { username: 'Pam5' },
-          { username: 'Pam6' },
-          { username: 'Pam7' },
-          { username: 'Pam8' },
-          { username: 'Pam9' },
-        ]}
-        maxHeight="350px"
-      />
-    </Grid.Row>
-  </>
-);
-
-const indicators = ['LOBBY', 'CHAT', 'GAMES'];
-const slides = [<LobbyInfo />, <Chat />, <GamesMenu />];
-
-const LobbyMobile = ({
-  theme,
-  windowDimensions,
-}: Pick<IStateProps, 'theme' | 'windowDimensions'>): ReactElement => {
-  const [activeIndex, setActiveIndex] = useState(0);
-  const sliderRef = useRef(document.createElement('div'));
-
-  return (
-    <div className="carousel">
-      <div className="carousel__nav">
-        {slides.map((_slide, i) => {
-          return (
-            <button
-              key={indicators[i]}
-              type="button"
-              onClick={(): void => {
-                sliderRef.current.scrollTo({
-                  left: sliderRef.current.scrollWidth * (i / 3),
-                  behavior: 'smooth',
-                });
-              }}
-              className={
-                i === activeIndex
-                  ? 'carousel__indicator active'
-                  : 'carousel__indicator'
-              }
-            >
-              {indicators[i]}
-            </button>
-          );
-        })}
-      </div>
-
-      <div
-        className="slider"
-        ref={sliderRef}
-        onScroll={(): void => {
-          const i = Math.round(
-            (sliderRef.current.scrollLeft * 3) / sliderRef.current.scrollWidth,
-          );
-          setActiveIndex(i);
-        }}
-      >
-        <div className="slider-content">
-          {slides.map(
-            (slide, i): ReactElement => {
-              return (
-                <div key={indicators[i]} className="slide">
-                  {slide}
-                </div>
-              );
-            },
-          )}
-        </div>
-      </div>
-
-      <style jsx>
-        {`
-          .carousel {
-            position: relative;
-            margin: 0 auto;
-            width: 100%;
-          }
-
-          .carousel__nav {
-            display: flex;
-            justify-content: center;
-            padding: 10px 0;
-          }
-
-          .carousel__indicator {
-            all: unset;
-            font-family: Montserrat-Bold;
-            background-color: ${theme.colors.BACKGROUND};
-            color: ${theme.colors.TEXT_GRAY};
-            font-size: 18px;
-            border-radius: 16px;
-            margin: 0 12px;
-            padding: 5px;
-          }
-
-          .carousel__indicator.active {
-            background-color: ${theme.colors.SLIDE_GOLD_BACKGROUND} !important;
-            color: ${theme.colors.GOLD} !important;
-          }
-
-          .slider {
-            position: relative;
-            margin: 0 auto;
-            height: 100%;
-            overflow-x: auto;
-            overflow-y: hidden;
-            scroll-snap-coordinate: 0 0;
-            scroll-snap-points-x: repeat(100%);
-            scroll-snap-type: x mandatory;
-            -webkit-overflow-scrolling: touch;
-          }
-
-          .slider-content {
-            height: 100%;
-            display: flex;
-          }
-
-          .slide {
-            height: 100%;
-            min-width: ${windowDimensions.width}px;
-            background-size: cover;
-            background-repeat: no-repeat;
-            background-position: center;
-            scroll-snap-align: start;
-          }
-
-          .slide.active {
-            display: block;
-          }
-        `}
-      </style>
-    </div>
-  );
-};
-
-const LobbyDesktop = (): ReactElement => (
-  <div className="container">
-    <Grid divided="vertically" padded className="main_grid">
-      <Grid.Column width={4}>
-        <Grid.Row className="logo_wrapper">
-          <Link href="/">
-            <a>
-              <img src="/common/logo.png" alt="logo" className="logo" />
-            </a>
-          </Link>
-        </Grid.Row>
-        <Grid.Row className="taako_wrapper">
-          <Taako />
-        </Grid.Row>
-        <Grid.Row>
-          <Announcements
-            announcements={[
-              {
-                id: '1',
-                timestamp: new Date(dateGenObj.next().value as number),
-                link: '/announcements/123',
-                text: 'New Patreon rewards!',
-              },
-              {
-                id: '2',
-                timestamp: new Date(dateGenObj.next().value as number),
-                link: '/announcements/123',
-                text: 'Morgana cannot lie anymore!',
-              },
-              {
-                id: '3',
-                timestamp: new Date(dateGenObj.next().value as number),
-                link: '/announcements/123',
-                text: 'Players will now be allowed to wear bows',
-              },
-            ]}
-          />
-        </Grid.Row>
-        <Grid.Row>
-          <OnlinePlayers
-            players={[
-              { username: 'ProNub' },
-              { username: 'Skies' },
-              { username: 'Pam' },
-              { username: 'Pam2' },
-              { username: 'Pam3' },
-              { username: 'Pam4' },
-              { username: 'Pam5' },
-              { username: 'Pam6' },
-              { username: 'Pam7' },
-              { username: 'Pam8' },
-              { username: 'Pam9' },
-            ]}
-            maxHeight="350px"
-          />
-        </Grid.Row>
-      </Grid.Column>
-
-      <Grid.Column width={12} className="center">
-        <Grid.Row className="navbar">
-          <Nav />
-        </Grid.Row>
-        <Grid.Row style={{ flex: 1 }}>
-          <Grid className="chat_games">
-            <Grid.Column width={11}>
-              <Chat />
-            </Grid.Column>
-            <Grid.Column width={5}>
-              <GamesMenu />
-            </Grid.Column>
-          </Grid>
-        </Grid.Row>
-      </Grid.Column>
-    </Grid>
-
-    <style jsx>
-      {`
-        .container {
-          padding: 30px;
-          z-index: -1;
-        }
-
-        .ui.grid {
-          margin-top: 0;
-        }
-
-        .logo {
-          max-width: 200px;
-          width: 100%;
-        }
-      `}
-    </style>
-  </div>
-);
-
-const mapStateToProps = (
-  state: RootState,
-): Pick<
-  IUserOptionsState & ISystemState,
-  'theme' | 'mobileView' | 'windowDimensions'
-> => ({
-  theme: state.userOptions.theme,
+const mapStateToProps = (state: RootState): IStateProps => ({
   mobileView: state.system.mobileView,
-  windowDimensions: state.system.windowDimensions,
 });
 
 export default connect(mapStateToProps, null)(Lobby);
